requirements.txt
customtkinter>=5.2.0
pdfplumber>=0.10.0
pypdf>=3.0.0
Pillow>=10.0.0
pymupdf>=1.23.0
pytesseract>=0.3.10
